New option to install setgid binary to call docker
Several improvements here.  First, I wanted a way to launch makerware
without belonging to the 'docker' group, without entering a sudo
password, and without making sudo passwordless.  The logical way to
accomplish this is to use a set-gid binary, but then it needs to be a
binary instead of a script.  I rewrote the script in C, and added the
`install-setgid` script to build it and run the sudo commands to set
the permissions.

While I was at it, I also have it generate a desktop menu file.

I discovered that if I use the --hostname option to make the container
match the host, I no longer need to open the xauth permissions to
allow other hosts, which should improve security, and used this in
both the new binary and the makerware-docker script.

Finally, I realized I wasn't allowing commandline arguments to be
passed to makerware, so I fixed that for both modes of execution.